+++ src/formio.embed.js
@@ -1,4 +1,4 @@
-import FormioForm from './formio.form.js';
+import { Formio } from './formio.full';
 let query = {};
 var scripts = document.getElementsByTagName('script');
 var thisScript = scripts[ scripts.length - 1 ];
@@ -10,4 +10,4 @@ queryString.replace(/\?/g, '&').split("&").forEach(function(item) {
   query[item.split("=")[0]] = item.split("=")[1] && decodeURIComponent(item.split("=")[1]);
 });
 query.styles = query.styles || (scriptSrc + '/formio.form.min.css');
-FormioForm.embed(query);
+Formio.embedForm(query);